requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
asn1crypto==1.3.0
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==12.0.3
django-easy-pdf==0.1.1
django-mail-templated==2.6.5
django-utils-six==2.0
django-widget-tweaks==1.4.8
fabric==2.5.0
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.9
invoke==1.4.1
paramiko==2.7.1
Pillow==8.1.0
pyasn1==0.4.8
pycparser==2.20
PyNaCl==1.3.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-decouple==3.3
pytz==2019.3
reportlab==3.5.59
six==1.14.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.5
